Linux fixes + Python 3 + PyQT5
I switched back to arch, so I fixed many of the linux issues, and made laptoplogger fully compatible with python3

Furthermore, since PyQt4 is nearing end of life, made the default gui toolkit PyQt5, with an auto-fallback on PyQt4.

I also implemented the logging routines myself for linux, meaning that we are back to being 100% MIT licensed!